[main] Let OperatorID::hash() return int.

+++ src/search/operator_id.h
@@ -45,7 +45,7 @@ public:
         return !(*this == other);
     }
 
-    size_t hash() const {
+    int hash() const {
         return index;
     }
 };
